final changes for pmax 4.4-Lite II release
SCCS-vsn: 8.3

+++ sys/pmax/dev/dtop.c
@@ -33,7 +33,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)dtop.c	8.2 (Berkeley) 11/30/93
+ *	@(#)dtop.c	8.3 (Berkeley) 06/02/95
  */
 
 /* 
@@ -305,7 +305,7 @@ dtopwrite(dev, uio, flag)
 /*ARGSUSED*/
 dtopioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
